feat: 백그라운드 스크립트 API 핸들러 리팩토링 및 추가
