+++ src/background/handlers/api/index.ts
@@ -1,0 +1,6 @@
+export { handleFoodDataFetch } from "./foodHandler";
+export { handleImageAnalysisFetch } from "./imageAnalysisHandler";
+export { handleOutlineInfoFetch } from "./outlineInfoHandler";
+export { handleCosmeticDataFetch } from "./cosmeticHandler";
+export { handleReviewSummaryFetch } from "./reviewSummaryHandler";
+export { handleHealthDataFetch } from "./healthDataHandler";
